--- a/src/Illuminate/Database/Attributes/SeedTask.php
+++ b/src/Illuminate/Database/Attributes/SeedTask.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Illuminate\Database\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_METHOD)]
+class SeedTask
+{
+    /**
+     * Create a new Seed Task instance.
+     */
+    public function __construct(public string $as = '')
+    {
+        //
+    }
+}