--- a/src/Illuminate/Database/Console/Seeds/SeedCommand.php
+++ b/src/Illuminate/Database/Console/Seeds/SeedCommand.php
@@ -7,9 +7,14 @@ use Illuminate\Console\ConfirmableTrait;
 use Illuminate\Console\Prohibitable;
 use Illuminate\Database\ConnectionResolverInterface as Resolver;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Seeder;
+use Illuminate\Filesystem\Filesystem;
+use RuntimeException;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
+use Throwable;
+use function json_encode;
 
 #[AsCommand(name: 'db:seed')]
 class SeedCommand extends Command
@@ -67,15 +72,108 @@ class SeedCommand extends Command
 
         $this->resolver->setDefaultConnection($this->getDatabase());
 
-        Model::unguarded(function () {
-            $this->getSeeder()->__invoke();
-        });
+        Model::unguarded($this->seed(...));
 
         if ($previousConnection) {
             $this->resolver->setDefaultConnection($previousConnection);
         }
 
         return 0;
+    }
+
+    /**
+     * Invoke the database seeder.
+     *
+     * @return void
+     */
+    protected function seed()
+    {
+        $seeder = $this->getSeeder();
+
+        $continueFilePath = $this->continueFilePath($seeder);
+
+        if ($this->option('continue')) {
+            if ($list = $this->retrieveContinueData($continueFilePath)) {
+                $this->components->info('Resuming from previous seed operation.');
+                Seeder::setContinue($list);
+            } else {
+                $this->components->warn('No previous seed operation found.');
+            }
+        }
+
+        if (! $seeder->useTransactions()) {
+            $this->components->warn('Transactions are disabled. Errors may yield incomplete records.');
+        }
+
+        try {
+            $seeder->__invoke();
+        } catch (Throwable $e) {
+            $this->storeContinueData($continueFilePath, Seeder::getContinue());
+
+            throw $e;
+        }
+
+        $this->deleteContinueData($continueFilePath);
+    }
+
+    /**
+     * Retrieve the "continue" data list from a JSON file as an array.
+     *
+     * @param  string  $path
+     * @return array
+     */
+    protected function retrieveContinueData($path)
+    {
+        if ($this->getLaravel()->bound(Filesystem::class)) {
+            $files = $this->getLaravel()->make(Filesystem::class);
+
+            return $files->exists($path) ? $files->json($path) : [];
+        }
+
+        throw new RuntimeException('Unable to retrieve continue data. Please install the "illuminate/filesystem" package to use the --continue option.');
+    }
+
+    /**
+     * Store the "continue" data list to a JSON file.
+     *
+     * @param  string  $path
+     * @param  array  $data
+     * @return void
+     */
+    protected function storeContinueData($path, array $data)
+    {
+        if ($this->getLaravel()->bound(Filesystem::class)) {
+            $files = $this->getLaravel()->make(Filesystem::class);
+
+            $files->ensureDirectoryExists(dirname($path));
+            $files->put($path, json_encode($data));
+        }
+    }
+
+    /**
+     * Returns the "continue" file path for the invoked seeder.
+     *
+     * @param \Illuminate\Database\Seeder $seeder
+     * @return string
+     */
+    protected function continueFilePath($seeder)
+    {
+        return $this->getLaravel()->storagePath(
+            'framework/database/seeder.' . str_replace('\\', '_', get_class($seeder)) . '.json'
+        );
+    }
+
+    /**
+     * Delete the "continue" file from the filesystem.
+     *
+     * @param  string  $path
+     * @return void
+     */
+    protected function deleteContinueData($path)
+    {
+        if ($this->getLaravel()->bound(Filesystem::class)) {
+            $this->getLaravel()->make(Filesystem::class)->delete($path);
+        }
     }
 
     /**
@@ -135,6 +233,7 @@ class SeedCommand extends Command
         return [
             ['class', null, InputOption::VALUE_OPTIONAL, 'The class name of the root seeder', 'Database\\Seeders\\DatabaseSeeder'],
             ['database', null, InputOption::VALUE_OPTIONAL, 'The database connection to seed'],
+            ['continue', null, InputOption::VALUE_OPTIONAL, 'Continue from a previous seed operation'],
             ['force', null, InputOption::VALUE_NONE, 'Force the operation to run when in production'],
         ];
     }