--- a/src/Illuminate/Database/Console/Seeds/SeedCommand.php
+++ b/src/Illuminate/Database/Console/Seeds/SeedCommand.php
@@ -101,7 +101,7 @@ class SeedCommand extends Command
             }
         }
 
-        if (! $seeder->useTransactions()) {
+        if (! method_exists($seeder, 'run') && ! $seeder->useTransactions()) {
             $this->components->warn('Transactions are disabled. Errors may yield incomplete records.');
         }