--- a/src/Illuminate/Database/Console/Seeds/stubs/seeder.stub
+++ b/src/Illuminate/Database/Console/Seeds/stubs/seeder.stub
@@ -8,9 +8,9 @@ use Illuminate\Database\Seeder;
 class {{ class }} extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Seed the database with records.
      */
-    public function run(): void
+    public function seed(): void
     {
         //
     }