--- a/src/Illuminate/Database/SeederSkipped.php
+++ b/src/Illuminate/Database/SeederSkipped.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Database;
+
+use RuntimeException;
+
+class SeederSkipped extends RuntimeException
+{
+    //
+}