… src/Illuminate/Database/composer.json
… tests/Database/DatabaseSeederTest.php
… tests/Database/SeedCommandTest.php