## What?

This PR supercharges the default seeder while keeping compatibility with classic seeders, enabling the following out-of-the-box:

- Run logic before/after the seeding is executed (great to skip or prepare records).
- Separate seeding into methods called automatically (seed tasks).
- Dependency Injection into each _seed task_.
- Skipping seeding completely or per _seed task_.
- Seeding errors tied to the seed task (to know _where_ it failed).
- Use a DB Transaction on each seed task to avoid orphaned records.
- Continue incomplete seeding.

## Why?

I do, more often than not, make seeders more handy in terms of programmatical seeding. For example, some tasks I do are:

- Use Dependency Injection for additional seeding tasks for a model.
- Skip the seeder if it detects some records already seeded.
- Wrapping seeding into transactions to speed up SQLite (try to seed 1000 records outside it!).
- Having different seeding sets.

The main idea that [I implemented on a project](https://larawiz.github.io/docs/model-properties/seeders.html) was to enable these tasks to run seamlessly, with less code, while separating seeding operations into more purposeful tasks than all-in-one method.

Once models and relations start to grow left and right, it becomes chaotic when seeders start to account for the multiple permutations and conditions, and all of these must be accounted by the developer with a class that doesn't help too much apart from calling other seeders.

## Give me the details

Okay. Let's start with each part from the easiest to understand nad the core of the seeder: the seed task.

### Seed tasks

Each seeder, when executed, will find all methods that start with the `seed` name or have the `#[SeedTask]` attribute (like PHPUnit does for tests), and call them (through the container) in the order these are declared in the class. There is no care for visibility, but `public` is preferred just in case the developer wants to call a seeder outside a command lifecycle.

```php
namespace Database\Seeders;

use Illuminate\Database\Attributes\SeedTask;
use Illuminate\Database\Seeder;
use Database\Factories\UserFactory;

class DatabaseSeeder extends Seeder
{
    public function run()
    {
        $this->call(UserSeeder::class);
    }
}

class UserSeeder extends Seeder
{
    public function seedNormalUsers(UserFactory $factory)
    {
        return $factory->count(5);
    }
    
    public function seedVipUsers(UserFactory $factory)
    {
        return $factory->count(3)->vip();
    }
 
    #[SeedTask(as: 'Seed users without authorization')]   
    public function bannedUsers(UserFactory $factory, $at = 'now')
    {
        return $factory->count(2)->banned($at);
    }
}
```

> [!TIP]
> 
> Yes, you can have a single seed task called `seed` and it will be called anyway. Great if you only have a single and simple seed task. Indeed, the updated Seeder stub comes with `seed()`.

If a method returns a `Factory` instance, an Eloquent Collection, or a Model, these will be automatically persisted in the database.

When executing the `db:seed` artisan command, the output will be something like this:

    php artisan db:seed
    
    INFO Seeding database.
    
    Database\Seeders\UserSeeder ...................................... RUNNING
    ↳ Seed normal users ................................................. DONE
    ↳ Seed vip users .................................................... DONE
    ↳ Seed users without authorization .................................. DONE
    Database\Seeders\UserSeeder .................................. 107 ms DONE
 
When the seeding fails, the seeding command will show the last seed task where it failed, making the seeding more informative, instead of seeing the Seeder name and figuring out _where_.

#### Calling with arguments

Currently, the `$this->call()` method accepts arguments, which are passed down to the `run()` method. When using the new seeder, the developer can set an array of arguments for each individual seed task by issuing their name as key and the arguments as an array.

```php
public function seedCommentTask()
{
    $this->call(CommentSeeder::class, [
        'bannedUsers' => ['at' => today()->subDay()]
    ]);
}
```

### Before, After, and onError.

A developer may need to set up some logic before and after a seeder runs with the `before()` and `after()` method, respectively. Maybe resetting a service status or configuration, or [skipping](#skipping-seed) the whole seeder. As with seed tasks, these are resolved by the Service Container. The `onError()` executes when a seed task throws an error, allowing the developer to do adjustments before throwing the exception, which van be great for reporting.

```php
/**
 * Execute logic before the seed tasks are executed 
 */
public function before()
{
    // Prepare some services...
}

/**
 * Execute logic after the seed tasks are executed.
 */
public function after()
{
    // Reset the services, remove artifacts...
}

public function onError($exception)
{
    // Remove some temporal files so we can try again...
}
```

I consider the above something more elegant than doing try-catch blocks everywhere.

### Skipping seed

There are two ways of skipping a seeder: completely or partially. Completely means skipping the seeder altogether, while partially means skipping only a given seed task.

Imagine we want to skip seeding the banned users if there is already one in existence. We only need to call `skip()` on the seed task, which will throw an exception that is safely caught by the seeder. The `skip()` method accepts a reason to skip, which is invaluable when seeding and figuring out why the seed task didn't complete or to avoid having more records that we need. 

```php
public function seedVipUsers()
{ 
    if (User::where('vip_at')->exists()) {
        $this->skip('There are already VIP users in the database');
    }
  
    // ...
}
```

When executing the `db:seed` artisan command, the output will be something like this, adding each seed task into the output:

    php artisan db:seed
    
    INFO Seeding database.
    
    Database\Seeders\UserSeeder ...................................... RUNNING
    ↳ Seed normal users ................................................. DONE
    ↳ Seed vip users ................................................. SKIPPED
    🛈 There are already VIP users in the database ...........................
    ↳ Seed banned users ................................................. DONE
    Database\Seeders\UserSeeder ................................... 95 ms DONE

To skip a seeder completely, the developer should call `skip()` inside the `before()` method.

```php
public function before()
{ 
    if (User::query()->exists()) {
        $this->skip('There are already users seeded in the database');
    }
  
    // ...
}
```

When executing the `db:seed` artisan command, the output will be something like this:

    php artisan db:seed
    
    INFO Seeding database.
    
    Database\Seeders\UserSeeder ...................................... RUNNING
    Database\Seeders\UserSeeder ...................................... SKIPPED
    🛈 There are already users seeded in the database ........................

### Transactions

The seeder **will wrap each seed task using a Database Transaction**. This not only makes seeding on SQLite out-of-the-box (file-based) more performant, but also to safely skip seeding without leaving orphaned or incomplete records. The latter will come in handy for [_resuming_ seeding](#stop--continue). 

The developer can disable transactions using the `useTransactions()` method of the seeder. If disabled, a warning will be shown since transactions are imperative for _safely_ continuing a previous failed seeding.  

```php
public function useTransactions()
{
    return false:
}
```

    php artisan db:seed
    
    INFO Seeding database.

    WARN Transactions are disabled. Errors may yield incomplete records.
    
    Database\Seeders\UserSeeder ...................................... RUNNING
    ↳ Seed normal users ................................................. DONE
    ↳ Seed vip users ................................................. SKIPPED
    🛈 There are already VIP users in the database ...........................
    ↳ Seed banned users ................................................. DONE
    Database\Seeders\UserSeeder ................................... 95 ms DONE

### Stop & Continue

One of my biggest problems when creating the seeders is that sometimes the seeding fails after a big seed. This means the developer needs to run the whole seeding operation again, especially of seeders depend on a previous one. This is why I implemented "Seed resuming", which can be enabled through the `--continue` option.

```shell
php artisan db:seed --continue
```

When the command executes, the list of seeders and seed tasks that ran is saved into a file (except when it completes). The file is named based on the seeder that ran, so these would have different lists saved:

```shell
php artisan db:seed --continue

php artisan db:seed --class=UserSeeder --continue
```

    storage/framework/database/seed.Database_Seeders_DatabaseSeeder.json
    
    storage/framework/database/seed.Database_Seeders_UserSeeder.json

The file is a JSON file like this, with the name of the class used to trigger the seeding and the list of called seeders and seed tasks.

```json
{
  "Database\\Seeders\\DatabaseSeeder": {
    "seedNormalUsers":  true,
    "seedVipUsers":  true,
    "seedBannedUsers":  true
  }
}
```

When the command receives the `--continue` option, it will check if there is a list, and load it into the Seeder. When the seeder calls the seeder's seed tasks, it will skip them as `CONTINUE` if these appear on the list. Once the whole seed completes successfully, the file is deleted.

    php artisan db:seed --continue
    
    INFO Seeding database.
    
    Database\Seeders\UserSeeder ...................................... RUNNING
    ↳ Seed normal users ............................................. CONTINUE
    ↳ Seed vip users ................................................ CONTINUE
    ↳ Seed banned users ................................................. DONE
    Database\Seeders\UserSeeder ................................... 95 ms DONE

What's imperative for the `--continue` to have transactions enabled. If a seed task fails, not wrapping the call with a transaction may create orphaned records.

> [!IMPORTANT]
> 
> Using the `--continue` option doesn't do anything on traditional seeders, or if the `illuminate/filesystem` package is not installed.

There is one issue for continuing an incomplete seeding operation: If the seed task was _skipped_, it counts as ran successfully and won't be invoked the next run.

## Retrocompatibility

I'm pretty confident this seeder is retro-compatible with the usual `run()`. Old projects do not need to change anything, new projects can get the updated stub with a single `seed()` method and be ready to work.