@DarkGhostHunter 2025-04-22
> Nicely done, however, I'm not sure if "task" is the ideal naming here—I might have gone with "step" or the like 🤔

I think you're right, since the new seeding behaves more like steps rather than async tasks. I'll let this concept marinate a while before making the changes.