@DarkGhostHunter 2025-04-23
> I would suggest releasing this as a package 👍

Any chances to rework the seeding system for Laravel 13?