@DarkGhostHunter 2025-04-27
Well, for those who expected a package out of this, I published it recently:

https://github.com/Laragear/Populate